+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/csv/JacksonCsvSerializer.kt
@@ -2,36 +2,36 @@ package com.windea.breezeframework.data.serializers.csv
 
 import com.fasterxml.jackson.dataformat.csv.*
 import java.io.*
+import java.lang.reflect.*
 
 object JacksonCsvSerializer : CsvSerializer {
 	@PublishedApi internal val mapper = CsvMapper()
 	
+	override fun <T> load(string: String, type: Type): T {
+		return mapper.readValue(string, mapper.typeFactory.constructType(type))
+	}
 	
-	override fun <T : Any> load(string: String, type: Class<T>): T {
+	override fun <T> load(file: File, type: Type): T {
+		return mapper.readValue(file, mapper.typeFactory.constructType(type))
+	}
+	
+	override fun <T> load(string: String, type: Class<T>): T {
 		return mapper.readValue(string, type)
 	}
 	
-	override fun <T : Any> load(file: File, type: Class<T>): T {
+	override fun <T> load(file: File, type: Class<T>): T {
 		return mapper.readValue(file, type)
 	}
 	
-	override fun <T : Any> load(reader: Reader, type: Class<T>): T {
-		return mapper.readValue(reader, type)
-	}
-	
-	override fun <T : Any> dump(data: T): String {
+	override fun <T> dump(data: T): String {
 		return mapper.writeValueAsString(data)
 	}
 	
-	override fun <T : Any> dump(data: T, file: File) {
+	override fun <T> dump(data: T, file: File) {
 		return mapper.writeValue(file, data)
-	}
-	
-	override fun <T : Any> dump(data: T, writer: Writer) {
-		return mapper.writeValue(writer, data)
 	}
 }
 
 object JacksonCsvSerializerConfig : CsvSerializerConfig {
-	inline operator fun invoke(builder: (CsvMapper) -> Unit) = builder(JacksonCsvSerializer.mapper)
+	inline fun configure(builder: (CsvMapper) -> Unit) = builder(JacksonCsvSerializer.mapper)
 }