+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/json/FastJsonSerializer.kt
@@ -2,29 +2,30 @@ package com.windea.breezeframework.data.serializers.json
 
 import com.alibaba.fastjson.*
 import java.io.*
+import java.lang.reflect.*
 
 object FastJsonSerializer : JsonSerializer {
-	override fun <T : Any> load(string: String, type: Class<T>): T {
+	override fun <T> load(string: String, type: Class<T>): T {
 		return JSON.parseObject(string, type)
 	}
 	
-	override fun <T : Any> load(file: File, type: Class<T>): T {
+	override fun <T> load(file: File, type: Class<T>): T {
 		return file.readText().let { JSON.parseObject(it, type) }
 	}
 	
-	override fun <T : Any> load(reader: Reader, type: Class<T>): T {
-		return reader.readText().let { JSON.parseObject(it, type) }
+	override fun <T> load(string: String, type: Type): T {
+		return JSON.parseObject(string, type)
 	}
 	
-	override fun <T : Any> dump(data: T): String {
+	override fun <T> load(file: File, type: Type): T {
+		return file.readText().let { JSON.parseObject(it, type) }
+	}
+	
+	override fun <T> dump(data: T): String {
 		return JSON.toJSONString(data, true)
 	}
 	
-	override fun <T : Any> dump(data: T, file: File) {
+	override fun <T> dump(data: T, file: File) {
 		return JSON.writeJSONString(file.writer(), data)
-	}
-	
-	override fun <T : Any> dump(data: T, writer: Writer) {
-		return JSON.writeJSONString(writer, data)
 	}
 }