+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/properties/JacksonPropertiesSerializer.kt
@@ -2,50 +2,49 @@ package com.windea.breezeframework.data.serializers.properties
 
 import com.fasterxml.jackson.dataformat.javaprop.*
 import java.io.*
+import java.lang.reflect.*
 import java.util.*
 
 object JacksonPropertiesSerializer : PropertiesSerializer {
 	@PublishedApi internal val mapper = JavaPropsMapper()
 	
-	
-	/**配置序列化选项。这个方法必须首先被调用。*/
-	fun configure(handler: (JavaPropsMapper) -> Unit): JacksonPropertiesSerializer {
-		return this.also { handler(mapper) }
-	}
-	
-	override fun <T : Any> load(string: String, type: Class<T>): T {
+	override fun <T> load(string: String, type: Class<T>): T {
 		return mapper.readValue(string, type)
 	}
 	
-	override fun <T : Any> load(file: File, type: Class<T>): T {
+	override fun <T> load(file: File, type: Class<T>): T {
 		return mapper.readValue(file, type)
-	}
-	
-	override fun <T : Any> load(reader: Reader, type: Class<T>): T {
-		return mapper.readValue(reader, type)
 	}
 	
 	override fun <T> load(properties: Properties, type: Class<T>): T {
 		return mapper.readPropertiesAs(properties, type)
 	}
 	
-	override fun <T : Any> dump(data: T): String {
+	override fun <T> load(string: String, type: Type): T {
+		return mapper.readValue(string, mapper.typeFactory.constructType(type))
+	}
+	
+	override fun <T> load(file: File, type: Type): T {
+		return mapper.readValue(file, mapper.typeFactory.constructType(type))
+	}
+	
+	override fun <T> load(properties: Properties, type: Type): T {
+		return mapper.readPropertiesAs(properties, mapper.typeFactory.constructType(type))
+	}
+	
+	override fun <T> dump(data: T): String {
 		return mapper.writeValueAsString(data)
 	}
 	
-	override fun <T : Any> dump(data: T, file: File) {
+	override fun <T> dump(data: T, file: File) {
 		return mapper.writeValue(file, data)
 	}
 	
-	override fun <T : Any> dump(data: T, writer: Writer) {
-		return mapper.writeValue(writer, data)
-	}
-	
-	override fun <T : Any> dump(data: T, properties: Properties) {
+	override fun <T> dump(data: T, properties: Properties) {
 		properties += mapper.writeValueAsProperties(data)
 	}
 }
 
 object JacksonPropertiesSerializerConfig : PropertiesSerializerConfig {
-	inline operator fun invoke(builder: (JavaPropsMapper) -> Unit) = builder(JacksonPropertiesSerializer.mapper)
+	inline fun configure(builder: (JavaPropsMapper) -> Unit) = builder(JacksonPropertiesSerializer.mapper)
 }