+++ breeze-reflect/src/main/kotlin/com/windea/breezeframework/reflect/extensions/CoreExtensions.kt
@@ -5,7 +5,7 @@ package com.windea.breezeframework.reflect.extensions
 import com.windea.breezeframework.core.annotations.api.*
 import kotlin.reflect.*
 
-/**判断指定Class是否在className中。*/
+/**判断指定名字的Class是否在classpath中。*/
 inline fun checkClassForName(className: String): Boolean {
 	return runCatching { Class.forName(className) }.onFailure { it.printStackTrace() }.isSuccess
 }