+++ breeze-spring-boot/src/main/kotlin/com/windea/breezeframework/springboot/core/CoreAutoConfiguration.kt
@@ -1,6 +1,5 @@
 package com.windea.breezeframework.springboot.core
 
-import com.windea.breezeframework.springboot.core.components.converters.*
 import com.windea.breezeframework.springboot.core.converters.*
 import org.springframework.context.annotation.*