+++ CHANGELOG.md
@@ -135,6 +135,7 @@
 * [X] breeze-core 提供字符串的逐行连接和逐行填充扩展。
 * [X] breeze-dsl MarkdownDsl 对于任何存在内联文本的元素，尽可能地实现`WrapContent`接口。
 * [X] breeze-core 添加`where`全局扩展。
+* [X] breeze-data Serializer 允许读取指定泛型类型的数据。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 
 ***